requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
numpy>=1.26.0
scipy>=1.12.0
Pillow>=10.2.0
python-dotenv>=1.0.0
playwright>=1.40.0
